+++ app/information/raw/RawClient.tsx
@@ -9,7 +9,7 @@ import { usePathname, useSearchParams } from "next/navigation";
 import { useState } from "react";
 import { RiArrowLeftSLine, RiArrowRightSLine } from "react-icons/ri";
 import Image from "next/image";
-import PicCeo from "@/public/img/page_top/ceo_top.jpg"
+import PicInfo_01 from "@/public/img/page_top/information_01.jpg"
 
 const MainList = [
   {
@@ -54,6 +54,20 @@ const RawClient: React.FC<RawProps> = ({ rawList, currentUser }) => {
 
   return (
     <section>
+      <figure className="w-full h-[200px]">
+        <div className="w-full h-full mx-auto relative">
+          <Image
+            src={PicInfo_01}
+            layout="fill"
+            objectFit="cover"
+            alt="item.title"
+            className="object-cover"
+          />
+          <div className="bg-neutral-900/50 absolute w-full h-full left-0 top-0 z-20 text-center text-white font-medium text-[36px] leading-[200px]">
+            {location}
+          </div>
+        </div>
+      </figure>
       <div id="headerNav">
         <div className="h-[40px] w-full bg-lightgray flex justify-center text-[13px]">
           <div className="w-full md:w-[1400px] flex justify-end pr-[20px]">
@@ -111,7 +125,7 @@ const RawClient: React.FC<RawProps> = ({ rawList, currentUser }) => {
                   <div className="w-[50px] text-center text-darkgray text-subtitle">
                     {item.id}
                   </div>
-                  <div className="ml-3">{item.title}</div>
+                  <div className="w-[1020px] ml-3 line-clamp-1">{item.title}</div>
                 </div>
                 <div className="w-[50px] text-darkgray ">
                   <RiArrowRightSLine className="text-[26px]" />
