+++ app/introduce/common/ceo/CeoClient.tsx
@@ -18,7 +18,7 @@ const MainList = [
     url: "#",
     sub: [
       { title: "CEO 인사말", url: "/introduce/common/ceo" },
-      { title: "비전/미션", url: "/introduce/common/vistion" },
+      { title: "비전/미션", url: "/introduce/common/vision" },
       { title: "연혁", url: "/introduce/common/history" },
       { title: "CI", url: "/introduce/common/ci" },
     ],
@@ -97,7 +97,7 @@ const CeoClient = () => {
           </div>
         </section>
 
-        <section className="md:py-[20px] md:pl-[40px] w-full flex flex-col justify-start items-start relative ">
+        <section className="md:py-[20px] md:pl-[40px] md:pr-[20px] w-full flex flex-col justify-start items-start relative ">
           <ContentTitle title={location}/>
           <ul className="flex md:hidden flex-wrap w-full py-[20px] px-[40px] text-[15px]">
             <li className="w-1/2 cursor-default">
