+++ app/mypage/payment/detail/DetailClient.tsx
@@ -59,8 +59,8 @@ const MainList = [
     title: "경력수첩 발급",
     url: "#",
     sub: [
-      { title: "경력수첩 발급", url: "/mypage/carrear/print" },
-      { title: "경력수첩 발급현황", url: "/mypage/carrear/sheet" },
+      { title: "경력수첩 발급", url: "/mypage/career/print" },
+      // { title: "경력수첩 발급현황", url: "/mypage/carrear/sheet" },
     ],
   },
   {
@@ -112,57 +112,74 @@ const DetailClient = () => {
           </div>
         </section>
 
-        <section className="p-[20px] w-full flex flex-col justify-start items-start">
+        <section className="py-[20px] pl-[40px] pr-[20px] w-full flex flex-col justify-start items-start">
           <ContentTitle title={location} center={true} />
           <div className="text-black w-full flex flex-col justify-center item-center">
-            <div className="w-full md:w-3/4 flex justify-center mx-auto mt-5">
+            <div className="w-full flex justify-center mx-auto mt-5">
               <p>
                 <span>회원</span>님의 납부내역입니다.<br/>
               </p>
             </div>
-            <ul className="w-full md:w-3/4 mx-auto border-t-2 border-neutral-700 mt-10 text-[14px] md:text-base">
+            <ul className="w-full mx-auto border-t-2 border-neutral-700 mt-10 text-[14px] md:text-base">
               <div className="w-full h-11 bg-lightgray flex text-center items-center font-medium border-b border-primary">
-                <div className="w-1/4 border-r border-gray-200">
+                <div className="w-[50px] h-full border-r border-gray-400 pt-[10px]">
+                  No.
+                </div>
+                <div className="w-1/4 h-full border-r border-gray-400 pt-[10px]">
                   금액
                 </div>
-                <div className="w-1/4 border-r border-gray-200">
+                <div className="w-1/4 h-full border-r border-gray-400 pt-[10px]">
                   납부상태
                 </div>
-                <div className="w-2/4">
+                <div className="w-2/4 h-full pt-[10px]">
                   일자
                 </div>
               </div>
               <li className="w-full h-11 flex text-center items-center border-b border-gray-400">
-                <div className="w-1/4 h-full border-r border-gray-200 flex pl-2 md:pl-10 items-center">
-                  ￦&nbsp;
+                <div className="w-[50px] h-full border-r border-gray-200 flex items-center justify-center">
+                  <span>
+                    1
+                  </span>
+                </div>
+                <div className="w-1/4 h-full border-r border-gray-200 flex items-center">
+                  &nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;￦&nbsp;
                   <span>
                     30,000
                   </span>
                 </div>
-                <div className="w-1/4 h-full border-r border-gray-200 flex pl-2 md:pl-10 items-center">
+                <div className="w-1/4 h-full border-r border-gray-200 flex items-center">
+                  &nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;
                   <span className="text-blue-600">
                     접수 대기
                   </span>
                 </div>
-                <div className="w-2/4 h-full border-r border-gray-200 flex pl-2 md:pl-10 items-center">
+                <div className="w-2/4 h-full border-r border-gray-200 flex items-center">
+                  &nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;
                   <span>
                     2023.12.15 17:51:01
                   </span>
                 </div>
               </li>
               <li className="w-full h-11 flex text-center items-center border-b border-gray-400">
-                <div className="w-1/4 h-full border-r border-gray-200 flex pl-2 md:pl-10 items-center">
-                  ￦&nbsp;
+                <div className="w-[50px] h-full border-r border-gray-200 flex items-center justify-center">
+                  <span>
+                    2
+                  </span>
+                </div>
+                <div className="w-1/4 h-full border-r border-gray-200 flex items-center">
+                  &nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;￦&nbsp;
                   <span>
                     500,000
                   </span>
                 </div>
-                <div className="w-1/4 h-full border-r border-gray-200 flex pl-2 md:pl-10 items-center">
+                <div className="w-1/4 h-full border-r border-gray-200 flex items-center">
+                  &nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;
                   <span className="text-red-600">
                     승인-입금완료
                   </span>
                 </div>
-                <div className="w-2/4 h-full border-r border-gray-200 flex pl-2 md:pl-10 items-center">
+                <div className="w-2/4 h-full border-r border-gray-200 flex items-center">
+                  &nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;
                   <span>
                     2024.10.25 07:54:01
                   </span>
