+++ app/notice/notice/NoticeClient.tsx
@@ -10,7 +10,7 @@ import { RiArrowLeftSLine, RiArrowRightSLine } from "react-icons/ri";
 
 import Pages from "@/components/pages";
 import Image from "next/image";
-import PicCeo from "@/public/img/page_top/ceo_top.jpg"
+import PicNotice_04 from "@/public/img/page_top/notice_04.jpg"
 
 const MainList = [
   {
@@ -67,7 +67,7 @@ const NoticeClient: React.FC<NoticeProps> = ({ currentUser, noticeList }) => {
       <figure className="w-full h-[200px]">
         <div className="w-full h-full mx-auto relative">
           <Image
-            src={PicCeo}
+            src={PicNotice_04}
             layout="fill"
             objectFit="cover"
             alt="item.title"
