+++ tailwind.config.js
@@ -9,7 +9,7 @@ module.exports = {
     extend: {
       colors: {
         // primary : '#003893',
-        primary : '#113678',
+        primary : '#003893',
         secondary : '#008ee5',
         lightgray : '#f5f5f5',
         darkgray : '#dbdbdb',
